backend/requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0

# Vision Pipeline
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0

# Face Analysis
insightface>=0.7.3
onnxruntime>=1.16.0
facenet-pytorch>=2.5.0
gdown>=4.7.0

# Pose Estimation (YOLO-Pose)
ultralytics>=8.0.0

# SMPLer-X Body Model
smplx>=0.1.28
trimesh>=4.0.0

# Utilities
python-multipart>=0.0.6
aiofiles>=23.0.0
scipy>=1.10.0
